[REFACTOR] utils/locators.py 로케이터 유지보수성 극대화를 위한 파일 분리 - PDP 로케이터
